review (move url parameter related code into file)
